blog_project/requirements.txt
Django==5.0
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
django-taggit==5.0.1
django-markdownx==4.0.7
Pillow>=10.3.0
markdown==3.5.2